Added activity tab

+++ mobile/src/app/friends/transaction-details/transaction-details.component.ts
@@ -10,14 +10,23 @@ import {
   IonTitle,
   IonToolbar, ModalController,
 } from '@ionic/angular/standalone';
-import { Friend, Transaction } from '../model';
+import { Friend, HistoryChangeType, Transaction } from '../model';
 import { NavParams } from '@ionic/angular';
 import { CurrencyPipe, DatePipe } from '@angular/common';
 import { DateFormatPipe } from '../../pipes/date-format.pipe';
 import { ShortenNamePipe } from '../../pipes/shorten-name.pipe';
 import { DefineExpenseComponent, SplitOptions } from '../../define-expense/define-expense.component';
 import { FriendsStore } from '../friends.store';
+import { HelperService } from '../../helper.service';
 
+const historyChangeType = {
+  [HistoryChangeType.DESCRIPTION]: "description",
+  [HistoryChangeType.TOTAL_AMOUNT]: "total amount",
+  [HistoryChangeType.CURRENCY]: "currency",
+  [HistoryChangeType.SPLIT_TYPE]: "split type",
+  [HistoryChangeType.DELETED]: "deleted",
+  [HistoryChangeType.TRANSACTION_DATE]: "transaction date"
+}
 @Component({
     selector: 'app-transaction-details',
     templateUrl: './transaction-details.component.html',
@@ -45,6 +54,7 @@ import { FriendsStore } from '../friends.store';
 export class TransactionDetailsComponent  implements OnInit {
   readonly SplitOptions = SplitOptions;
   readonly friendsStore = inject(FriendsStore);
+  readonly helperService = inject(HelperService);
   readonly modalCtrl = inject(ModalController);
   readonly nav = inject(IonNav);
   readonly transaction = input.required<Transaction>();
@@ -54,28 +64,28 @@ export class TransactionDetailsComponent  implements OnInit {
 
   ngOnInit() {}
 
-  ionViewWillEnter() {
-    this.friendsStore.setSelectedFriend(this.friend());
-  }
-
-  ionViewWillLeave() {
-    this.friendsStore.setSelectedFriend(null);
-  }
-
   async deleteTransaction() {
-    await this.friendsStore.deleteTransaction(this.transaction());
-    this.nav.pop();
+    const response = await this.helperService.showConfirmAlert()
+    if (response.role === 'confirm') {
+      try {
+        await this.friendsStore.deleteTransaction(this.transaction());
+        this.nav.pop();
+      } catch (e) {}
+    }
   }
 
   async updateTransaction() {
     const modal = await this.modalCtrl.create({
       component: DefineExpenseComponent,
       componentProps: {
-        friend: { name: this.transaction().friendName },
+        friend: this.friend(),
         isUpdating: true,
         transaction: this.transaction(),
       }
     })
     modal.present();
   }
+
+  protected readonly historyChangeType = historyChangeType;
+  protected readonly HistoryChangeType = HistoryChangeType;
 }
